Made Kernel.java as interface, extracted class into BaseKernel.java
