+++ src/main/java/de/dmi3y/behaiv/kernel/Kernel.java
@@ -1,95 +1,37 @@
 package de.dmi3y.behaiv.kernel;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import de.dmi3y.behaiv.storage.BehaivStorage;
 import de.dmi3y.behaiv.tools.Pair;
 
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.FileReader;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
-public abstract class Kernel {
+public interface Kernel {
+    void setId(String id);
 
-    protected String id;
-    protected Long treshold = 10L;
-    protected ObjectMapper objectMapper;
-    protected boolean partialFitAllowed = false;
-    protected boolean alwaysKeepData = true;
+    boolean isEmpty();
 
-    public Kernel(String id) {
-        this.id = id;
-        objectMapper = new ObjectMapper();
-    }
+    void fit(List<Pair<List<Double>, String>> data);
 
+    void fit();
 
-    public void setId(String id) {
-        this.id = id;
-    }
+    void setTreshold(Long treshold);
 
+    boolean readyToPredict();
 
-    //list<features>, label
-    protected List<Pair<List<Double>, String>> data = new ArrayList<>();
+    void update(List<Pair<List<Double>, String>> data);
 
+    boolean isPartialFitAllowed();
 
-    public abstract boolean isEmpty();
+    void updateSingle(List<Double> features, String label);
 
-    public abstract void fit(List<Pair<List<Double>, String>> data);
+    String predictOne(List<Double> features);
 
-    public void fit() {
-        fit(this.data);
-    }
+    boolean isAlwaysKeepData();
 
-    public void setTreshold(Long treshold) {
-        this.treshold = treshold;
-    }
+    void setAlwaysKeepData(boolean alwaysKeepData);
 
-    public boolean readyToPredict() {
-        return data.size() > treshold;
-    }
+    void save(BehaivStorage storage) throws IOException;
 
-    public void update(List<Pair<List<Double>, String>> data) {
-    }
-
-    public boolean isPartialFitAllowed() {
-        return partialFitAllowed;
-    }
-
-    public void updateSingle(List<Double> features, String label) {
-        data.add(new Pair<>(features, label));
-    }
-
-    public abstract String predictOne(List<Double> features);
-
-    public boolean isAlwaysKeepData() {
-        return alwaysKeepData;
-    }
-
-    public void setAlwaysKeepData(boolean alwaysKeepData) {
-        this.alwaysKeepData = alwaysKeepData;
-    }
-
-    public void save(BehaivStorage storage) throws IOException {
-
-        try (final BufferedWriter writer = new BufferedWriter(new FileWriter(storage.getDataFile(id)))) {
-            writer.write(objectMapper.writeValueAsString(data));
-        }
-    }
-
-    public void restore(BehaivStorage storage) throws IOException {
-        final TypeReference<List<Pair<List<Double>, String>>> typeReference = new TypeReference<List<Pair<List<Double>, String>>>() {
-        };
-        try (final BufferedReader reader = new BufferedReader(new FileReader(storage.getDataFile(id)))) {
-            final String content = reader.readLine();
-            if (content == null || content.isEmpty()) {
-                data = new ArrayList<>();
-            } else {
-                data = objectMapper.readValue(content, typeReference);
-            }
-        }
-    }
+    void restore(BehaivStorage storage) throws IOException;
 }
